Set claims appropriately on the id_token
The old version of this code would have the effect of never setting any of the name, email or role claims on the identity token. This change fixes that, and makes this controller consistent with similar lines in the Refresh Flow, Implicit Flow and Authorization Code Flow example authorization controllers.