Allow newer sdk versions
I've just updated visual studio to 17.5. And I got only the 7.0.200 sdk which made it impossible to compile the solution after checking out. I am sure it was not meant to be fixed to this version of the runtime, certainly because lots of people will try this out. An easy workaround for this is to add the rollforward option, as described here:

https://learn.microsoft.com/en-us/dotnet/core/tools/global-json#rollforward